+++ nata-monkey/src/main/java/com/nata/AllRunner.java
@@ -11,30 +11,16 @@ import com.nata.monkeys.RandomMonkey;
  */
 public class AllRunner {
     public static void main(String[] args) {
-
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";
 
         //Set Monkey settings
         AdbDevice device = new AdbDevice();
-        RandomMonkey randomMonkey = new RandomMonkey(pkg,act,device);
-        QLearningMonkey qLearningMonkey= new QLearningMonkey(pkg,act,device);
-
-        for(int i = 0; i  < 3; i++){
-            //Start run monkey
-            randomMonkey.play();
-
-            //Do report
-            randomMonkey.report();
-        }
-
-        for(int i = 0 ; i < 3 ; i++){
-            //Start run monkey
-            qLearningMonkey.play();
-
-            //Do report
-            qLearningMonkey.report();
-        }
-
+        RandomMonkey randomMonkey = new RandomMonkey(pkg, act, device);
+        QLearningMonkey qLearningMonkey = new QLearningMonkey(pkg, act, device);
+        randomMonkey.play();
+        randomMonkey.report();
+        qLearningMonkey.play();
+        qLearningMonkey.report();
     }
 }