+++ nata-monkey/src/main/java/com/nata/AndroidKeyCode.java
@@ -11,11 +11,11 @@ public class AndroidKeyCode {
     public static int HOME = 3;
     public static int MENU = 82;
     public static int VOLUME_UP = 24;
-    public static  int VOLUME_DOWN = 25;
-    public static  int SPACE = 62;
+    public static int VOLUME_DOWN = 25;
+    public static int SPACE = 62;
     public static int BACKSPACE = 67;
     public static int ENTER = 66;
-    public static  int MOVE_HOME = 122;
+    public static int MOVE_HOME = 122;
     public static int MOVE_END = 123;
     public static int CLEAR = 28;
     public static int HIDE_KEYBOARD = 111;