+++ nata-monkey/src/main/java/com/nata/RMRunner.java
@@ -10,23 +10,14 @@ import com.nata.monkeys.RandomMonkey;
  */
 public class RMRunner {
     public static void main(String[] args) {
-        //Info about the apk to test
-//        String pkg = "com.zhihu.android";
-//        String act = ".app.ui.activity.MainActivity";
+//      String pkg = "com.zhihu.android";
+//      String act = ".app.ui.activity.MainActivity";
 
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";
-        //Set Monkey settings
         AdbDevice device = new AdbDevice();
-        RandomMonkey randomMonkey = new RandomMonkey(pkg,act,device);
-
-        for(int i = 0; i  < 3; i++){
-            //Start run monkey
-            randomMonkey.play();
-
-            //Do report
-            randomMonkey.report();
-        }
-
+        RandomMonkey randomMonkey = new RandomMonkey(pkg, act, device);
+        randomMonkey.play();
+        randomMonkey.report();
     }
 }